fix pc plot, GAP py files fix session structure

+++ lineup_app/views.py
@@ -40,14 +40,17 @@ login_manager.init_app(app)
 
 # users dictionary to login to the app
 # temporary solution, must be placed outside of the code (database or xls)
-users = { # can add more users if needed
-    'admin':{'password':'kpo'},
-}
+users_json=os.path.join(uploader_dirname,r"temp\users.json")
+users = json.load(open(users_json))
+
 
 # Driver to be used to query data from EC. This is different on different computers
 db_driver="Oracle in OraClient11g"
 #db_driver="Oracle in OraClient11g_32_bit"
 
+
+# using session_json instead of flask session due to cookie size limitation
+session_json={}
 
 
 """ LOGIN ROUTINES (using flask-login library) ================================================================ """
@@ -87,8 +90,10 @@ def logout():
 # if not, redirect back to login page
 @app.route('/login', methods=['GET', 'POST'])
 def login():
+    page_active={"load_pcs":"","load_state":"","setup":"","live":"","results":""}
     if request.method == 'POST':
         username = request.form.get('username')
+        # print(username,users)
         if username in users:
             if request.form.get('password') == users[username]['password']:
                 user = User()
@@ -98,9 +103,7 @@ def login():
             else:
                 flash("Incorrect")
         else:
-            return render_template('login.html')
-
-    page_active={"load_pcs":"","load_state":"","setup":"","live":"","results":""}
+            return render_template('login.html',page_active=page_active)
     return render_template('login.html',page_active=page_active)
 
 # error handing in case of incorrect or no login, redirects to login page
@@ -116,7 +119,7 @@ def index():
     session["state"]={} # initialize state if no state
     session["state"]=loaded_state # save loaded data to state
     session.modified=True
-    print(session["state"]["fb_data"]["streams"]["constraints"])
+    # print(session["state"]["fb_data"]["streams"]["constraints"])
     page_active={"load_pcs":"","load_state":"","setup":"","live":"","results":""}
     return render_template('index.html',page_active=page_active)
 
@@ -154,10 +157,28 @@ def results():
     # get MAPs from "Deliverability,xlsx" file
     well_maps=xl_setup.read_maps()
 
+    # get AFs to apply on well level
+    afs={
+        "kpc":{
+            "af_gas":session["state"]["fb_data"]["wells"]["kpc"]["af_gas"],
+            "af_oil":session["state"]["fb_data"]["wells"]["kpc"]["af_oil"]
+        },
+        "u3":{
+            "af_gas":session["state"]["fb_data"]["wells"]["u3"]["af_gas"],
+            "af_oil":session["state"]["fb_data"]["wells"]["u3"]["af_oil"]
+        },
+        "u2":{
+            "af_gas":session["state"]["fb_data"]["wells"]["u2"]["af_gas"],
+            "af_oil":session["state"]["fb_data"]["wells"]["u2"]["af_oil"]
+        }
+    }
+
+    # session["state"]["fb_data"]["wells"]
+
     # get well results from GAP
     #------------------------------------------------------------------------------
-    # pc_data=ggr.get_all_unit_pc(well_details,session["state"]["wells"])
-    pc_data=xlggr.xl_get_all_unit_pc(well_details,session["state"]["well_state"])
+    pc_data=ggr.get_all_unit_pc(well_details,session["state"]["well_state"],afs)
+    # pc_data=xlggr.xl_get_all_unit_pc(well_details,session["state"]["well_state"])
     #------------------------------------------------------------------------------
 
     # a dictionary for mapping unit index and name
@@ -220,27 +241,27 @@ def setup():
 
     if "well_state" in session["state"]:
         #------------------------------------------------------------------------------
-        #st.set_unit_routes(well_details,session["state"]["well_state"]) # set well routes as per state
-        xlst.xl_set_unit_routes(well_details,session["state"]["well_state"]) # set well routes as per state
+        st.set_unit_routes(well_details,session["state"]["well_state"]) # set well routes as per state
+        # xlst.xl_set_unit_routes(well_details,session["state"]["well_state"]) # set well routes as per state
         #------------------------------------------------------------------------------
 
     if "sep" in session["state"]["unit_data"]:
         #------------------------------------------------------------------------------
-        #st.set_sep_pres(session["state"]["sep"]) # set separator pressure as per state
-        xlst.xl_set_sep_pres(session["state"]["unit_data"]["sep"]) # set separator pressure as per state
+        st.set_sep_pres(session["state"]["unit_data"]["sep"]) # set separator pressure as per state
+        # xlst.xl_set_sep_pres(session["state"]["unit_data"]["sep"]) # set separator pressure as per state
         #------------------------------------------------------------------------------
 
 
     data=session["state"]
 
     #------------------------------------------------------------------------------
-    #data["well_data"]=st.get_all_well_data(well_details) # get well data from GAP such as GOR, limits and current routes
-    data["well_data"]=xlst.xl_get_all_well_data(well_details) # get well data from GAP such as GOR, limits and current routes
+    data["well_data"]=st.get_all_well_data(well_details) # get well data from GAP such as GOR, limits and current routes
+    # data["well_data"]=xlst.xl_get_all_well_data(well_details) # get well data from GAP such as GOR, limits and current routes
     #------------------------------------------------------------------------------
 
     #------------------------------------------------------------------------------
-    # data["sep"]=st.get_sep_pres() # get separator pressure if state doesn't exist
-    data["unit_data"]["sep"]=xlst.xl_get_sep_pres() # get separator pressure if state doesn't exist
+    data["unit_data"]["sep"]=st.get_sep_pres() # get separator pressure if state doesn't exist
+    # data["unit_data"]["sep"]=xlst.xl_get_sep_pres() # get separator pressure if state doesn't exist
     #------------------------------------------------------------------------------
 
 
@@ -282,7 +303,12 @@ def clearstate():
 def savestate():
     session.pop('state', None) # clears state in session
     session["state"]=request.json # set state to the data passed from AJAX call in setup page javascript
+    print(session["state"]["well_state"])
     session.modified=True
+
+    # session_jsonfile=os.path.join(uploader_dirname,r"temp\session.json")
+    # json.dump(request.json, open(session_jsonfile, 'w'))
+
     return "None"
 """========================================================================================="""
 
@@ -382,7 +408,7 @@ def savestate_loaded():
 def gap_calc():
     print('received start command!')
     #------------------------------------------------------------------------------
-    # post_opt_state=gob.run_optimization(session["state"]) # pass state to GAP to make calculations
+    post_opt_state=gob.run_optimization(session["state"]) # pass state to GAP to make calculations
     #skip calc when xl is used
     # post_opt_state=xlgob.xl_run_optimization(session["state"]) # pass state to GAP to make calculations
     #------------------------------------------------------------------------------
@@ -420,12 +446,13 @@ def save_2ref():
     return "None"
 """========================================================================================="""
 
-""" DELETE REFERENCE JSON FILE==================================================== """
+""" REMOVE REFERENCE JSON FILE==================================================== """
 
 def delete_refcase():
     dirname, filename = os.path.split(os.path.abspath(__file__))
     json_fullpath_ref=os.path.join(dirname,r"temp\results_ref_case.json")
-    os.remove(json_fullpath_ref)
+    if os.path.isfile(json_fullpath_ref):
+        os.remove(json_fullpath_ref)
     return "None"
 """========================================================================================="""
